+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - generator for `:module` value of the `type` check
 - `allow_nil` has `false` value by default
 - `map` generator without `length` opts provided produces [min: 0, max: 10] maps
+- with no `type` check specified, but with `inner` a param is `:map` by default
 
 ## [0.1.2] - 2018.12.05